chore: Rename push-notifications to notifications
