+++ projects/Mallard/src/download-edition/download-via-notification.ts
@@ -2,7 +2,7 @@ import ApolloClient from 'apollo-client'
 import { matchSummmaryToKey } from 'src/helpers/files'
 import { imageForScreenSize } from 'src/helpers/screen'
 import { getIssueSummary } from 'src/hooks/use-issue-summary'
-import { pushTracking } from 'src/push-notifications/push-tracking'
+import { pushTracking } from 'src/notifications/push-tracking'
 import { errorService } from 'src/services/errors'
 import { Feature } from '../../../Apps/common/src/logging'
 import { downloadAndUnzipIssue } from './download-and-unzip'
