+++ projects/Mallard/src/helpers/diagnostics.ts
@@ -11,7 +11,7 @@ import { FSPaths } from 'src/paths'
 import { runActionSheet } from './action-sheet'
 import { getFileList } from './files'
 import { locale } from './locale'
-import { getDiagnosticPushTracking } from '../push-notifications/push-tracking'
+import { getDiagnosticPushTracking } from '../notifications/push-tracking'
 import { isInBeta } from './release-stream'
 import { imageForScreenSize } from './screen'
 import { iapReceiptCache, userDataCache, pushRegisteredTokens } from './storage'
